// frontend/src/api/models/repairSortBy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * backend-node
 * OpenAPI spec version: 1.0.0
 */

export type RepairSortBy = typeof RepairSortBy[keyof typeof RepairSortBy];


export const RepairSortBy = {
  createdAt: 'createdAt',
  costoTotale: 'costoTotale',
  stato: 'stato',
} as const;